+++ .github/workflows/CI-Rust.yml
@@ -70,21 +70,19 @@ jobs:
           toolchain: stable
           components: rustfmt,clippy
       - run: |
-          cargo fmt --check
-          cargo clippy --no-default-features --features server-only -- -D warnings
           sudo apt-get -y update
-          sudo apt-get -y install musl-tools musl-dev perl
+          sudo apt-get -y install musl-tools musl-dev perl libdbus-1-dev pkg-config libxcb1-dev libxcb-render0-dev libxcb-shape0-dev libxcb-xfixes0-dev
           rustup target add x86_64-unknown-linux-musl
-          cargo build --no-default-features --features server-only --verbose --release --target x86_64-unknown-linux-musl
-          sudo apt-get install -y libdbus-1-dev pkg-config libxcb1-dev libxcb-render0-dev libxcb-shape0-dev libxcb-xfixes0-dev
-          cargo test
+          cargo fmt --check
           cargo clippy -- -D warnings
+          cargo test
+          cargo build --verbose --release --target x86_64-unknown-linux-musl --package clip-sync-server
           cargo build --verbose --release
       - uses: actions/upload-artifact@v3
         with:
           retention-days: 14
           name: clip-sync-server-x86_64-linux-musl
-          path: target/x86_64-unknown-linux-musl/release/clip-sync
+          path: target/x86_64-unknown-linux-musl/release/clip-sync-server
       - uses: actions/upload-artifact@v3
         with:
           retention-days: 14
