+++ .vscode/settings.json
@@ -1,4 +1,5 @@
 {
+  "rust-analyzer.check.allTargets": false,
   "editor.formatOnSave": true,
   "[toml]": {
     "editor.formatOnSave": false
