+++ client-interface/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "client-interface"
+version = "0.6.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = { workspace = true }
+log = { workspace = true }
+png = { workspace = true }
+serde = { workspace = true, features = ["derive"] }
+gethostname = { workspace = true }
+chrono = { workspace = true, optional = true}
+
+[features]
+websocket = [
+    "chrono"
+]
