+++ mqtt-client/src/lib.rs
@@ -3,7 +3,7 @@ use log::{debug, warn};
 use rumqttc::{AsyncClient, EventLoop, MqttOptions, QoS};
 use serde::Deserialize;
 
-use crate::{clipboard_handler, ClipboardRecord, ClipboardSink, ClipboardSource};
+use client_interface::{ClipSyncClient, ClipboardRecord, ClipboardSink, ClipboardSource};
 
 #[derive(Debug, Clone, Default, Deserialize)]
 #[serde(deny_unknown_fields, rename_all = "kebab-case")]
@@ -16,7 +16,7 @@ pub struct MqttClientConfig {
     pub mqtt_client_id: Option<String>,
 }
 
-struct MqttSubscriber {
+pub struct MqttSubscriber {
     eventloop: EventLoop,
     device_id: String,
 }
@@ -55,7 +55,7 @@ impl ClipboardSource for MqttSubscriber {
     }
 }
 
-struct MqttPublisher {
+pub struct MqttPublisher {
     client: AsyncClient,
     topic: String,
 }
@@ -78,30 +78,39 @@ impl ClipboardSink for MqttPublisher {
     }
 }
 
-pub async fn clip_sync_svc(args: MqttClientConfig) -> anyhow::Result<()> {
-    let sender_id = args.mqtt_client_id.unwrap_or(
-        gethostname()
-            .into_string()
-            .unwrap_or(random_string::generate(12, "abcdefghijklmnopqrstuvwxyz")),
-    );
-    let mut options = MqttOptions::new(
-        sender_id.clone(),
-        args.mqtt_server_addr,
-        args.mqtt_server_port,
-    );
+pub struct MqttClipSyncClient;
 
-    if args.mqtt_username.is_some() || args.mqtt_password.is_some() {
-        options.set_credentials(
-            args.mqtt_username.unwrap_or_default(),
-            args.mqtt_password.unwrap_or_default(),
+impl ClipSyncClient for MqttClipSyncClient {
+    type Config = MqttClientConfig;
+
+    #[allow(refining_impl_trait)]
+    async fn connect(
+        args: Self::Config,
+    ) -> anyhow::Result<(String, MqttSubscriber, MqttPublisher)> {
+        let sender_id = args.mqtt_client_id.unwrap_or(
+            gethostname()
+                .into_string()
+                .unwrap_or(random_string::generate(12, "abcdefghijklmnopqrstuvwxyz")),
         );
+        let mut options = MqttOptions::new(
+            sender_id.clone(),
+            args.mqtt_server_addr,
+            args.mqtt_server_port,
+        );
+
+        if args.mqtt_username.is_some() || args.mqtt_password.is_some() {
+            options.set_credentials(
+                args.mqtt_username.unwrap_or_default(),
+                args.mqtt_password.unwrap_or_default(),
+            );
+        }
+
+        let topic = args.mqtt_topic.unwrap_or("clipboard".to_string());
+        let (client, eventloop) = AsyncClient::new(options, 10);
+        client.subscribe(topic.clone(), QoS::AtLeastOnce).await?;
+
+        let sink = MqttPublisher::new(client.clone(), topic.clone());
+        let source = MqttSubscriber::new(eventloop, sender_id.clone());
+        Ok((sender_id, source, sink))
     }
-
-    let topic = args.mqtt_topic.unwrap_or("clipboard".to_string());
-    let (client, eventloop) = AsyncClient::new(options, 10);
-    client.subscribe(topic.clone(), QoS::AtLeastOnce).await?;
-
-    let sink = MqttPublisher::new(client.clone(), topic.clone());
-    let source = MqttSubscriber::new(eventloop, sender_id.clone());
-    clipboard_handler::start(sender_id, source, sink).await
 }
