+++ websocket-client/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "websocket-client"
+version = "0.6.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = { workspace = true }
+log = { workspace = true }
+clap = { workspace = true, features = ["derive"] }
+tokio = { workspace = true, features = ["sync", "rt", "rt-multi-thread", "fs"] }
+serde = { workspace = true, features = ["derive"] }
+serde_json = { workspace = true }
+futures = { workspace = true }
+futures-util = { workspace = true }
+gethostname = { workspace = true }
+tokio-tungstenite = { workspace = true }
+url = { workspace = true }
+reqwest = { workspace = true, features = ["json", "multipart"] }
+random-string = { workspace = true }
+
+client-interface = { workspace = true, features = ["websocket"] }
