+++ websocket-server/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "websocket-server"
+version = "0.6.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = { workspace = true }
+log = { workspace = true }
+clap = { workspace = true, features = ["derive"] }
+tokio = { workspace = true, features = ["sync", "rt", "rt-multi-thread", "fs"] }
+serde = { workspace = true, features = ["derive"] }
+serde_json = { workspace = true }
+poem = { workspace = true, features = ["websocket", "rustls", "static-files", "multipart"] }
+tantivy = { workspace = true }
+chrono = { workspace = true }
+futures-util = { workspace = true }
+
+client-interface = { workspace = true, features = ["websocket"] }
