+++ websocket-server/src/auth.rs
